+++ src/services/repo/hooks.spec.js
@@ -5,8 +5,6 @@ import { setupServer } from 'msw/node'
 import { MemoryRouter, Route } from 'react-router-dom'
 
 import {
-  MEASUREMENT_TYPE,
-  useActivateFlagMeasurements,
   useEraseRepoContent,
   useRepo,
   useRepoBackfilled,
@@ -263,54 +261,6 @@ describe('useRepoBackfilled', () => {
         }
         await waitFor(() =>
           expect(result.current.data).toEqual(expectedResponse)
-        )
-      })
-    })
-  })
-})
-
-describe('useActivateFlagMeasurements', () => {
-  function setup() {
-    server.use(
-      graphql.mutation('ActivateMeasurements', (req, res, ctx) => {
-        return res(
-          ctx.status(200),
-          ctx.data({
-            activateMeasurements: null,
-          })
-        )
-      })
-    )
-  }
-
-  describe('when called', () => {
-    beforeEach(() => {
-      setup()
-    })
-
-    describe('When success', () => {
-      it('returns expected output', async () => {
-        const { result } = renderHook(
-          () =>
-            useActivateFlagMeasurements({
-              provider: 'gh',
-              owner: 'dancer',
-              repo: 'bassuras',
-              measurementType: MEASUREMENT_TYPE.FLAG_COVERAGE,
-            }),
-          {
-            wrapper: wrapper(),
-          }
-        )
-
-        result.current.mutate()
-
-        await waitFor(() =>
-          expect(result.current.data).toEqual({
-            data: {
-              activateMeasurements: null,
-            },
-          })
         )
       })
     })
